requirements.txt
# 核心依赖
requests>=2.28.0
dingtalk-stream>=0.24.2
python-dotenv>=0.19.0
sseclient-py>=1.7.2

# 钉钉官方SDK
alibabacloud-dingtalk>=2.2.27
alibabacloud-tea-openapi>=0.4.0
alibabacloud-tea-util>=0.3.13

# 工具类
pytz>=2022.1
colorama>=0.4.4  # Windows终端颜色支持
typing-extensions>=4.1.1

# SSL和安全
pyOpenSSL>=20.0.1
certifi>=2021.10.8
urllib3>=1.26.0 